Refactor "application events" into UI classes
This means that there are less global state and it's easier to test what
we care about, as well as being easier to plug in another interface if
the need arises. It's especially good since it decouples the interface
from the application and Application becomes nothing more than an
orchestrator.